demo/requirements.txt
gradio
ultralytics
opencv-python
pillow
pandas
numpy
